refactor: Menu 가격 범위 검증 Menu 클래스로 이동
