+++ src/test/java/kitchenpos/MenuAcceptanceTest.java
@@ -55,27 +55,6 @@ class MenuAcceptanceTest extends AcceptanceTest {
         }
 
         @Test
-        void 메뉴는_0_이상의_가격을_가진다() {
-            final MenuGroup menuGroup = 일식();
-            final Long menuGroupId = 메뉴_그룹_생성_요청하고_아이디_반환(menuGroup);
-            final Long productId = 상품_생성_요청하고_아이디_반환(스키야키());
-
-            final MenuProduct menuProduct = new MenuProduct();
-            menuProduct.setProductId(productId);
-            menuProduct.setQuantity(1L);
-
-            final Menu menu = new Menu();
-            menu.setName("스키야키");
-            menu.setPrice(BigDecimal.valueOf(-1));
-            menu.setMenuGroupId(menuGroupId);
-            menu.setMenuProducts(List.of(menuProduct));
-
-            final ExtractableResponse<Response> response = 메뉴_생성_요청(menu);
-
-            assertThat(response.statusCode()).isEqualTo(INTERNAL_SERVER_ERROR.value());
-        }
-
-        @Test
         void 메뉴는_반드시_메뉴_그룹에_속해야_한다() {
             final Product product = 스키야키();
             final Long productId = 상품_생성_요청하고_아이디_반환(product);
